Implement Grocery List Management API with CRUD operations and API usage tracking
- Added GroceryListController for handling HTTP requests related to grocery lists.
- Created GroceryListDAO interface and its implementation for database interactions.
- Implemented methods for retrieving, saving, updating, and deleting grocery lists and items.
- Introduced GroceryListRequest and GroceryListResponse classes for request and response handling.
- Integrated ApiUsageDAO to track API usage for each operation.
- Developed ApiUsageDAO and its implementation for managing API usage logs.
- Added UserProfile and MealPlanRequest classes for user-related data handling.
- Enhanced RecipeDAO and RecipeDAOImpl for managing recipes and meal plans.
- Implemented security measures with ApiKeyFilter and SecurityConfig for API key validation.
- Added CORS configuration to allow cross-origin requests.

+++ demo/src/main/java/com/shubhodip/nutrisift/user/User.java
@@ -3,9 +3,10 @@ package com.shubhodip.nutrisift.user;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
+import java.util.concurrent.atomic.LongAccumulator;
 
 public class User {
-    private int id;
+    private long id;
     private Timestamp time;
     private Date date;
     private String userid;
@@ -22,11 +23,11 @@ public class User {
     }
 
     // Getters and setters
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
